+++ src/main/java/com/tsp/new_tsp_front/api/portfolio/domain/FrontPortFolioDTO.java
@@ -26,7 +26,7 @@ public class FrontPortFolioDTO extends NewCommonDTO {
     private Integer rnum;
 
     @ApiModelProperty(required = true, value = "idx", hidden = true)
-    private Integer idx;
+    private Long idx;
 
     @NotEmpty(message = "제목은 필수입니다.")
     @ApiModelProperty(required = true, value = "title")
